Client Stats IOCTL sets correct Tx-bytes/frames (2.4G)
Tx-bytes parameter is filled in with too large value for 2.4GHz WiFi
clients when throughput test is run.
Tx-frames parameter is fixed in the same way.
The issue is caused by the legacy code which dates back to QSDK 2.0
and other very old driver bases with their quirks.
This is no longer necessary on newer QSDK releases.

Fixes: 89fbc4bc9008 ("Release 1.4.0.1")

Signed-off-by: Comcast-Serhiy-Chumak <serhiy_chumak@comcast.com>